Correct compile error

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -3,6 +3,7 @@
 
 /**
  * array_iterator - execute function on each array's element
+ * @array: array of elements
  * @size: size of array
  * @action: pointer to function to be used
  *
@@ -13,7 +14,7 @@ void array_iterator(int *array, size_t size, void (*action)(int))
 
 	if (array && size && action)
 	{
-		for (i = 0; i < size; i++)
+		for (i = 0; i < (int) size; i++)
 		{
 			action(array[i]);
 		}
